+++ src/main/java/com/erapulus/server/dto/BuildingRequestDto.java
@@ -24,10 +24,10 @@ public class BuildingRequestDto {
 
     @NotNull
     @JsonProperty("latitude")
-    private Long latitude;
+    private Double latitude;
 
     @NotNull
     @JsonProperty("longitude")
-    private Integer longitude;
+    private Double longitude;
 
 }
